requirements.txt
psycopg[binary]==3.2.9
pymodbus==3.8.3
python-dotenv==1.0.1
